Webtech-Assignment2/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
frozenlist==1.4.1
google-api-core==2.15.0
google-auth==2.27.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
packaging==23.2
platformdirs==4.2.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
urllib3==2.1.0
virtualenv==20.25.0
Werkzeug==3.0.1
yarl==1.9.4
